[3주차 과제 RacingCar] RacingZone 뜯어고치기 전 커밋.
