+++ src/main/kotlin/Main.kt
@@ -1,19 +1,6 @@
-import racingcar.RacingZone
-import racingcar.RacingCars
-import racingcar.view.InputView
-import racingcar.view.ResultView
-
-fun main() {
-    val inputView = InputView()
-    val numOfCar = inputView.getNumOfCar()
-    val numOfTry = inputView.getNumOfTry()
-
-    val racingCars = RacingCars()
-    RacingCars().prepareCar(numOfCar)
-
-    val racingZone = RacingZone()
-    val resultOfRace = racingZone.startRaceAndGetResult(racingCars.getCurrentCarList(), numOfTry)
-
-    val resultView = ResultView()
-    resultView.showResult(resultOfRace)
-}
+//fun main() {
+//    val keyboardInputString = readLine() ?: throw IllegalArgumentException("입력값은 null일 수 없습니다.")
+//    val splitList = SplitString().splitString(keyboardInputString)
+//    val resultValue = StringCalculator().calculate(splitList)
+//    print("계산값은 ${resultValue}입니다.")
+//}
