+++ src/main/kotlin/racingcar/RacingZone.kt
@@ -9,7 +9,9 @@ class RacingZone {
         var resultOfRace = ""
         for (i in 0 until numOfTry) {
             resultOfRace += (runOneCycle(carList)) + (LINE_SPACE)
+            print("전체 차가 한 번의 레이스를 시도한 결과: $resultOfRace")
         }
+        print("전체 레이스 끝났을 때: $resultOfRace")
         return resultOfRace
     }
 
@@ -20,7 +22,9 @@ class RacingZone {
             val currentCarStepList = racingCar.getCarStepList()
             val currentCarStepString = currentCarStepList.joinToString(separator = "")
             resultOfOneCycle += currentCarStepString + LINE_SPACE
+            print("한 번씩 돌때 차량 한대: $resultOfOneCycle")
         }
+        print("한 번씩 돌때 전체: $resultOfOneCycle")
         return resultOfOneCycle
     }
 }
