+++ src/main/kotlin/stringcalculator/Main.kt
@@ -1,0 +1,8 @@
+package stringcalculator
+
+fun main() {
+    val keyboardInputString = readLine() ?: throw IllegalArgumentException("입력값은 null일 수 없습니다.")
+    val splitList = SplitString().splitString(keyboardInputString)
+    val resultValue = StringCalculator().calculate(splitList)
+    print("계산값은 ${resultValue}입니다.")
+}
